state and behaviour tests in one class
